requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-import-export==4.1.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
pillow==11.0.0
PyJWT==2.9.0
sqlparse==0.5.1
tablib==3.5.0
